Refactor : 서비스 소개 관련 레이아웃 수정, 스타일 단일화
하얀 공백을 중간에 삽입하였던 기존의 스타일을 변경하고, 통일된 형식의 디자인을 채용하였습니다.
