+++ src/constants/ServiceInfoData.js
@@ -6,7 +6,7 @@ export const ServiceInfoData = [
   {
     title: "모두에게, 모두가.",
     desc: "도심을 푸르게 만드는 옥상 녹화는\n홍수 방지, 열섬 효과 차단, 건물 내구성 강화와 같이\n모두의 공간으로 환원됩니다.",
-    bottomText: "개인의 옥상에서 모두의 공간으로.",
+    bottomText: "개인의 옥상에서 모두의 공간으로, 우리는 변화합니다.",
     img: service1,
   },
   {
@@ -17,7 +17,7 @@ export const ServiceInfoData = [
   },
   {
     title: "녹초가 된 치유공간.",
-    desc: "치유받는 공간을 만들기 위해 우리는 옥상을 녹화합니다.\n안정감을 통한 건전한 스트레스 절감을 위해.",
+    desc: "모두가 치유받는 공간을 만들기 위해,\n안정감을 통한 건전한 스트레스 절감을 위해.\n우리는 도시의 옥상을 푸르게 녹화합니다.",
     bottomText: "모두의 삶 속에 녹색이 스며들 때까지, 우리는 초록을 만듭니다.",
     img: service3,
   },
